Generated XML files.
Ran raster.sh to generate the state required to run xml.sh.
Ran xml.sh to generate iconpack.xml and drawable.xml
Added lines of the XML from the IconRequest to appfilter.xml and ran
sort.sh to sort and format the file.

Removed mpv icon in favour of the one now in upstream.
Adjusted appfilter.xml in preparation for merging in upstream changes.