removed the two out of three apps
could only find two of the three requested apps here so removed those two. 
'Pixel Search' app was already missing in the requests.txt so not sure if it was already uploaded or not. Also not sure what to do if I have already made one.
Will there be a face-off thing between the two icons or will it be first-come-first-serve?