Update appfilter.xml to fix typo with Canvas Teacher
The component ID was off (it's "activit_ies_", plural, not activit_y_, singular).  This is necessary because it's not actually displaying the Arcticon icon on my phone (and hopefully this is why)

Also, I went back to using my icon.  This part I'm not going to insist on - if you prefer the other one (the same one that's being used for the Canvas Student app) that's fine by me.